fix: Cache recipe filters at startup instead of querying DB per request
